Update user model to generate and verify password
