+++ app/__init__.py
@@ -2,9 +2,10 @@ from flask import Flask
 from flask_graphql import GraphQLView
 from flask_sqlalchemy import SQLAlchemy
 from config import config
-from . import schema
+from app import graphql
 
 db = SQLAlchemy()
+
 
 def create_app(config_name):
     app = Flask(__name__)
@@ -17,7 +18,7 @@ def create_app(config_name):
         '/graphql',
         view_func=GraphQLView.as_view(
             'graphql',
-            schema=schema,
+            schema=graphql.schema,
             graphiql=True,
         )
     )
