+++ app/graphql/__init__.py
@@ -1,0 +1,1 @@
+from .schema import schema
