+++ app/graphql/schema.py
@@ -1,9 +1,11 @@
 import graphene
 
+
 class Query(graphene.ObjectType):
     hello = graphene.String(description='A typical hello world')
 
     def resolve_hello(self, info):
         return 'World'
+
 
 schema = graphene.Schema(query=Query)
