+++ app/models/__init__.py
@@ -1,0 +1,2 @@
+from .user import User
+from .contact import Contact
