+++ app/models/user.py
@@ -1,0 +1,27 @@
+from werkzeug.security import generate_password_hash, check_password_hash
+from sqlalchemy.dialects.postgresql import UUID
+from app import db
+
+
+class User(db.Model):
+    __tablename__ = 'users'
+
+    # Properties
+    id = db.Column(UUID, primary_key=True)
+    first_name = db.Column(db.String, nullable=False)
+    last_name = db.Column(db.String, nullable=False)
+    password_hash = db.Column(db.String)
+
+    # Relations
+    contacts = db.relationship('Contact', backref='user')
+
+    @property
+    def password(self):
+        raise AttributeError('Password is not a readable attribute')
+
+    @password.setter
+    def password(self, password):
+        self.password_hash = generate_password_hash(password)
+
+    def verify_password(self, password):
+        return check_password_hash(self.password_hash, password)
