+++ manage.py
@@ -1,8 +1,9 @@
 from app import create_app, db
-from flask_migrate import Migrate, upgrade
+from flask_migrate import Migrate
 
 app = create_app('development')
 migrate = Migrate(app, db)
+
 
 @app.shell_context_processor
 def make_shell_context():
